+++ Castle/src/Entities/Entity.java
@@ -8,30 +8,30 @@ public interface Entity {
 
     public void update(int delta);
 
-    public void setLocation(double x, double y);
+    public void setLocation(float x, float y);
 
-    public void setX(double x);
+    public void setX(float x);
 
-    public void setY(double y);
+    public void setY(float y);
 
-    public void setWidth(double width);
+    public void setWidth(float width);
 
-    public void setHeight(double height);
+    public void setHeight(float height);
 
-    public double getX();
+    public float getX();
     
-    public double getCenterX();
+    public float getCenterX();
 
-    public double getY();
+    public float getY();
     
-    public double getCenterY();
+    public float getCenterY();
 
-    public double getHeight();
+    public float getHeight();
 
-    public double getWidth();
+    public float getWidth();
 
     public boolean intersects(Entity other);
     
-    public double getIntersectAt(Entity other, Entity primary);
+    public float getIntersectAt(Entity other, Entity primary);
 
 }
